// src/app/core/models/JobTypes/jobtype-level.model.ts
export class JobtypeLevel {
    constructor() {

    }
    JobTypeLevelId: number;
    JobTypeId: number;
    JobTypeLevelUid: string;
    JobTypeLevelName: string;
    JobTypeName: string;
    IsActive: boolean;
    IsDeleted: boolean;
    OrgId?: number;
    CreatedBy?: number;
    CreatedOn: string;
    LastModifiedBy?: number;
    LastModifiedOn: string;
    OrgName : string;
}